Handle pytest tests that take a long time or many statuses
When tests take more than a minute, the summary includes a test time in
wall clock format which wasn't handled by the previous regex. Also, if
the summary line is very long because there are many different kinds of
test statuses, the regex might not match because there aren't as many ==
separators. It's probably still possible to get fewer, but we need to
match some to get reasonably few false matches, so it's probably about
right now.