Add the --gather-analysis and --comment options to analyzepr
This retrieves failed test results for one or more PRs from a single CI
(origin) at a time and stores the stats in a file.  Once results from
all CI services have been received, the --comment option can be used to
write a GitHub PR comment with a summary of the failure analysis if a
test failure occurs.