Fix pytype log parsing on GHA
The log filter classes were switched to derive directly from TextIO in
commit 98ac2785 which caused the seek method to go to the empty one in
the base class instead of the real one on the embedded file member.
Since the TextIO was there only for typing purposes, eliminate it and
use a (very ugly) synthetic type class instead. Unfortunately, I don't
know of a cleaner way to keep typing on these classes that satisfies
pytype. Even this doesn't even work with pytype on Python 3.11, so
disable that version in the CI.

Fix MsBuildLog to reset its in_msbuild flag when the file is rewound.

Fix the filters' readline() to all support the size argument.