Switch to a private API to get logs on Cirrus
The public v1.1 API truncates the logs, so use a private API to get the
full log instead. The private API also provides raw logs whereas the
public one wraps them in JSON. Since this API is private, leave the
capability in the code to switch back to the public API at a moment's
notice in case the private one stops working (which happened at least
once in the last year or so).