Track the PR commit tested when gathering failures
This allows the user to be notified of the commit, as well as when it is
inconsistent across CI services.

This changes the pr_gather_path persistent state file format, so that
file should be deleted after upgrading to avoid a crash. The experimental
upgrade function in prdef could be run to upgrade the objects instead.