Disable the user configuration file when running tests
We have no control over the contents of that file, so ignore it while
testing. Remove the call to unittest.main() which was only
inconsistently available and rely on the test gathering that unittest
and pytest automatically do. Add calls to the superclass's setUp() to
make sure nothing is skipped.