Don't confuse arch metadata with hostarch in curlparse
arch is the architecture of the machine on which the tests are being
run. hostarch is the configuration for that machine, which might be a 32
bit version of the former. It turns out buildinfo.host.cpu is redundant,
as hostarch already is extracted from buildinfo.host, so ignore the
former and don't use it to set anything. Add a cross-compiled test case
that doesn't really test this specific difference, but it does at least
test a different target specification.