Handle logs files with corrupt character encoding
It can happen that a binary log file is dumped within the log to be
parsed, which can cause character set decode errors when reading it in
(especially in UTF-8). If this occurs, replace any bad characters with
backslash escape sequences in order to successfully read the remainder
of the file.