Process all files in log zipfile for GHA
No longer skip log files in the root directory in preference to those in
subdirectories. This was an optimization to avoid parsing the same log
files twice, but GHA seems to have stopped including the subdirectory
copies about 2025-06-30 (although it's not entirely clear if that was a
GHA change or a change to curl's job definitions).  This new format
also means the cistep metadata is no longer available.